requirements.txt
grpcio
grpcio-tools
flask
cs50
Flask-Session
requests
tkinter
